requirements.txt
Flask
python-dotenv
SQLAlchemy
PyMySQL